+++ Enderlook.Pools/src/Instance/SafeObjectPool.cs
@@ -119,9 +119,9 @@ public sealed class SafeObjectPool<T> : ObjectPool<T>
     }
 
     /// <summary>
-    /// Determines if the reserve pool is allowed to grow and shrink given its usage.
+    /// Determines if the reserve pool is not allowed to grow nor shrink given its usage.
     /// </summary>
-    public bool IsReserveDynamic { get; init; } = true;
+    public bool IsReserveFixed { get; init; } = true;
 
     /// <summary>
     /// If this value is not <see langword="null"/>, the callback will be executed on each element which is free from the pool.<br/>
@@ -568,7 +568,7 @@ public sealed class SafeObjectPool<T> : ObjectPool<T>
         Array reserve_ = Utils.NullExchange(ref reserve);
         int oldReserveCount = reserveCount;
         int oldReserveLength = reserve_.Length;
-        bool isReserveDynamic = IsReserveDynamic;
+        bool isReserveFixed = IsReserveFixed;
         // Under high pressure, we don't wait time to trim, so we remove all objects in reserve.
         if (reserveTrimPercentage == 1)
         {
@@ -593,7 +593,7 @@ public sealed class SafeObjectPool<T> : ObjectPool<T>
                 newReserveCount = Math.Max(oldReserveCount - toRemove, 0);
 
                 // Since the global reserve has a dynamic size, we shrink the reserve if it gets too small.
-                if (isReserveDynamic && oldReserveLength / newReserveCount >= ReserveShrinkFactorToStart)
+                if (!isReserveFixed && oldReserveLength / newReserveCount >= ReserveShrinkFactorToStart)
                 {
                     Debug.Assert(ReserveShrinkFactorToStart >= ReserveShrinkFactor);
                     newReserveLength = Math.Max(oldReserveLength / ReserveShrinkFactor, itemsLength);
@@ -609,7 +609,7 @@ public sealed class SafeObjectPool<T> : ObjectPool<T>
             }
         }
 
-        newReserveLength = isReserveDynamic ? Math.Max(newReserveLength, Math.Min(oldReserveLength, itemsLength)) : oldReserveLength;
+        newReserveLength = isReserveFixed ? oldReserveLength : Math.Max(newReserveLength, Math.Min(oldReserveLength, itemsLength));
         Debug.Assert(newReserveLength <= oldReserveLength);
         Debug.Assert(newReserveLength >= newReserveCount);